+++ js/screens/greeting.js
